+++ R/available-packages.R
@@ -549,6 +549,6 @@ renv_available_packages_cellar <- function(type, project = NULL) {
   })
 
   # bind into data.frame for lookup
-  lbind(records)
+  bind(records)
 
 }