+++ tests/testthat/test-utils.R
@@ -11,63 +11,6 @@ test_that("common utils work as expected", {
     expect_error(git())
 })
 
-test_that("lbind handles named lists", {
-
-  data <- list(
-    alpha = list(A = 1, B = 2),
-    beta  = list(A = 3, B = 4),
-    gamma = list(A = 5, B = 6)
-  )
-
-  actual <- lbind(data)
-  expected <- data.frame(
-    Index = names(data),
-    A = c(1, 3, 5),
-    B = c(2, 4, 6),
-    stringsAsFactors = FALSE
-  )
-
-  expect_equal(actual, expected)
-
-})
-
-test_that("lbind warns on name collision", {
-  data <- list(alpha = list(Index = 1), beta = list(Index = 2))
-  expect_error(lbind(data))
-})
-
-test_that("lbind() handles data.frames with potentially different names", {
-
-  data <- list(
-    a = data.frame(A = 1, B = 2),
-    b = data.frame(A = 1, C = 3)
-  )
-
-  bound <- lbind(data)
-  expected <- data.frame(
-    Index = c("a", "b"),
-    A     = c(1, 1),
-    B     = c(2, NA),
-    C     = c(NA, 3),
-    stringsAsFactors = FALSE
-  )
-
-  expect_identical(bound, expected)
-
-})
-
-test_that("lbind() preserves order where possible", {
-
-  data <- list(
-    a = data.frame(A = 1,        C = 3),
-    b = data.frame(A = 1, B = 2, C = 3)
-  )
-
-  bound <- lbind(data)
-  expect_equal(names(bound), c("Index", "A", "B", "C"))
-
-})
-
 test_that("versions are compared as expected", {
 
   expect_equal(renv_version_compare("0.1.0", "0.2.0"), -1L)
@@ -172,3 +115,4 @@ test_that("recursing() reports if we're recursing", {
 
 
 })
+